+++ mayan/apps/lock_manager/tests/test_backends.py
@@ -1,5 +1,3 @@
-import time
-
 from unittest import skip
 from django.test import override_settings
 from django.utils.module_loading import import_string
@@ -41,7 +39,8 @@ class FileLockTestCase(BaseTestCase):
         with self.assertRaises(expected_exception=LockError):
             self.locking_backend.acquire_lock(name=TEST_LOCK_1)
 
-        time.sleep(1.01)
+        self._test_delay(seconds=1.01)
+
         # lock_1 not release but has expired, should not raise LockError
         lock_2 = self.locking_backend.acquire_lock(name=TEST_LOCK_1)
 
@@ -54,7 +53,9 @@ class FileLockTestCase(BaseTestCase):
 
     def test_release_expired(self):
         lock_1 = self.locking_backend.acquire_lock(name=TEST_LOCK_1, timeout=1)
-        time.sleep(1.01)
+
+        self._test_delay(seconds=1.01)
+
         lock_1.release()
         # No exception is raised even though the lock has expired.
         # The logic is that checking for expired locks during release is
@@ -63,7 +64,9 @@ class FileLockTestCase(BaseTestCase):
 
     def test_release_expired_reaquired(self):
         self.locking_backend.acquire_lock(name=TEST_LOCK_1, timeout=1)
-        time.sleep(1.01)
+
+        self._test_delay(seconds=1.01)
+
         # TEST_LOCK_1 is expired so trying to acquire it should not return an
         # error.
         lock_2 = self.locking_backend.acquire_lock(name=TEST_LOCK_1, timeout=1)