+++ HISTORY.rst
@@ -181,3 +181,4 @@
   apps to define their own default ``SearchModel``.
 - Update MySQL Docker image from version 5.7 to 8.0. PostreSQL image from version
   10.14 to 10.15. Redis image from version 5.0 to 6.0.
+- Move time delays from test and into its own test mixin. Remove MySQL test delays.